# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DRIP. API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  username: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  username: string;
  isPremium: boolean;
  totalScans: number;
  highestScore: number;
  avgScore: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface CategoryScore {
  score: number;
  detail: string;
}

export interface OccasionMatch {
  casualHangout: number;
  dateNight: number;
  workOffice: number;
  formalEvent: number;
  streetwearFlex: number;
}

export interface UpgradeRecommendation {
  swap: string;
  with: string;
  predictedScoreIncrease: number;
  searchQuery: string;
  priceRange: string;
}

export interface CelebrityMatch {
  name: string;
  matchPercentage: number;
  description: string;
}

export type ScanResultScores = {
  fit: CategoryScore;
  color: CategoryScore;
  styleCoherence: CategoryScore;
  trendAlignment: CategoryScore;
  signatureFactor: CategoryScore;
};

export interface ScanResult {
  id: number;
  userId: number;
  imagePath: string;
  overallScore: number;
  scores: ScanResultScores;
  styleArchetype: string;
  archetypeDescription: string;
  vibeTags: string[];
  bestElement: string;
  weakestElement: string;
  occasionMatch: OccasionMatch;
  upgrades: UpgradeRecommendation[];
  similarQueries: string[];
  colorPalette: string[];
  season: string;
  celebrityMatch: CelebrityMatch;
  genderDetected: string;
  createdAt: string;
}

export interface ScanSummary {
  id: number;
  imagePath: string;
  overallScore: number;
  styleArchetype: string;
  createdAt: string;
}

export interface UserScanStats {
  totalScans: number;
  avgScore: number;
  highestScore: number;
  /** @nullable */
  highestScoreDate?: string | null;
}

export interface ScanHistoryResponse {
  scans: ScanSummary[];
  total: number;
  page: number;
  totalPages: number;
  stats: UserScanStats;
}

export interface SimilarOutfit {
  id: number;
  imageUrl: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  styleTag?: string | null;
}

/**
 * @nullable
 */
export type StyleDnaResponseAvgScores = { [key: string]: unknown } | null;

export type StyleDnaResponseStyleEvolutionItem = {
  archetype: string;
  date: string;
  score: number;
};

export type StyleDnaResponseScoreTrendItem = {
  date: string;
  score: number;
};

export interface StyleDnaResponse {
  locked: boolean;
  scansNeeded: number;
  /** @nullable */
  dominantArchetype?: string | null;
  /** @nullable */
  archetypePercentage?: number | null;
  topColors?: string[];
  /** @nullable */
  avgScores?: StyleDnaResponseAvgScores;
  totalOutfitsScanned: number;
  bestOutfit?: ScanSummary;
  worstOutfit?: ScanSummary;
  styleEvolution?: StyleDnaResponseStyleEvolutionItem[];
  scoreTrend?: StyleDnaResponseScoreTrendItem[];
}

export type StyleReportResponseScoreTrend =
  (typeof StyleReportResponseScoreTrend)[keyof typeof StyleReportResponseScoreTrend];

export const StyleReportResponseScoreTrend = {
  up: "up",
  down: "down",
  stable: "stable",
} as const;

export interface StyleReportResponse {
  available: boolean;
  totalOutfits: number;
  avgScore: number;
  scoreTrend: StyleReportResponseScoreTrend;
  /** @nullable */
  mostWornStyle?: string | null;
  bestOutfit?: ScanSummary;
  topColorPalette?: string[];
  /** @nullable */
  styleInOneWord?: string | null;
}

export interface LeaderboardEntry {
  id: number;
  username: string;
  overallScore: number;
  styleArchetype: string;
  imagePath: string;
  createdAt: string;
}

export interface AppStats {
  totalScans: number;
  totalUsers: number;
  avgScore: number;
  topArchetype: string;
}

export type CreateScanBody = {
  image: Blob;
};

export type GetScanHistoryParams = {
  page?: number;
  limit?: number;
  sort?: GetScanHistorySort;
};

export type GetScanHistorySort =
  (typeof GetScanHistorySort)[keyof typeof GetScanHistorySort];

export const GetScanHistorySort = {
  date_desc: "date_desc",
  date_asc: "date_asc",
  score_desc: "score_desc",
  score_asc: "score_asc",
} as const;

export type GetLeaderboardParams = {
  period?: GetLeaderboardPeriod;
  limit?: number;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  daily: "daily",
  weekly: "weekly",
  all_time: "all_time",
} as const;