# lib/api-zod/src/generated/types/scanHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DRIP. API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanSummary } from "./scanSummary";
import type { UserScanStats } from "./userScanStats";

export interface ScanHistoryResponse {
  scans: ScanSummary[];
  total: number;
  page: number;
  totalPages: number;
  stats: UserScanStats;
}