# lib/api-zod/src/generated/types/upgradeRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DRIP. API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpgradeRecommendation {
  swap: string;
  with: string;
  predictedScoreIncrease: number;
  searchQuery: string;
  priceRange: string;
}