# lib/api-zod/src/generated/types/userScanStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DRIP. API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserScanStats {
  totalScans: number;
  avgScore: number;
  highestScore: number;
  /** @nullable */
  highestScoreDate?: Date | null;
}